Detect batch/block differences

Detect batch/block differences within analytical runs for each ionisation mode.

detectBatchDiff(x, by = "block", pthresh = 0.05)

# S4 method for Binalysis
detectBatchDiff(x, by = "block", pthresh = 0.05)

# S4 method for MetaboProfile
detectBatchDiff(x, by = "block", pthresh = 0.05)

Arguments

x
object of class Binalysis or MetaboProfile

by
info class column to use for batch information

pthresh
p-value threshold for significance

Value

If no differences between batches are found then NULL is returned. If significant differences are found then a tibble is returned containing the ANOVA results for each ionisation mode and showing whether batch correction is needed.

Details

Analysis of Variance (ANOVA) is used to detect differences in total ion count (TIC) averages between batches/blocks.

Examples

## Retrieve file paths and sample information for example data
files <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')[1:2]

info <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')[1:2,]

## Perform spectral binning
analysis <- binneR::binneRlyse(files, 
                               info, 
                               parameters = binneR::detectParameters(files))
#> 
#> Attaching package: ‘purrr’
#> The following object is masked from ‘package:metaboMisc’:
#> 
#>     reduce
#> binneR v2.6.2 Wed Nov 17 09:52:25 2021
#> ________________________________________________________________________________
#> Scans: 5:14 
#> ________________________________________________________________________________
#> Reading raw data
#> Gathering bins
#> Removing single scan events
#> Averaging intensities across scans
#> Calculating bin metrics
#> Calculating accurate m/z
#> Building intensity matrix
#> Gathering file headers
#> 
#> Completed! [2.6S]

## Detect batch differences
batch_diff <- detectBatchDiff(analysis)
#> Batches with < 3 replicates removed: "3", "5"
#> Only 1 batch detected, skipping detection

## Display batch diffferences
batch_diff
#> NULL