Detect miss injections

Detect miss injected samples.

detectMissInjections(x, idx = "fileOrder")

# S4 method for Binalysis
detectMissInjections(x, idx = "fileOrder")

# S4 method for MetaboProfile
detectMissInjections(x, idx = "fileOrder")

Arguments

x

object of class Binalysis or MetaboProfile

idx

sample information column to use for sample indexes

Value

A list containing the name of sample information column used to index the miss injections and a vector of miss injection indexes.

Details

Samples with a total ion count (TIC) below 1.5 times the inter-quartile range are detected as miss injections.

Examples

## Retrieve file paths and sample information for example data
files <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')[1:2]

info <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')[1:2,]

## Perform spectral binning
analysis <- binneR::binneRlyse(files, 
                               info, 
                               parameters = binneR::detectParameters(files))
#> binneR v2.6.2 Thu Dec 16 13:30:30 2021
#> ________________________________________________________________________________
#> Scans: 5:14 
#> ________________________________________________________________________________
#> Reading raw data
#> Gathering bins
#> Removing single scan events
#> Averaging intensities across scans
#> Calculating bin metrics
#> Calculating accurate m/z
#> Building intensity matrix
#> Gathering file headers
#> 
#> Completed! [1.1S]

## Detect miss injections
miss_injections <- detectMissInjections(analysis)

## Display detected miss injections
miss_injections$missInjections
#> numeric(0)