Pre-treatment of multiple ionisation modes

Pre-treat both positive and negative ionisation modes for Binalysis and MetaboProfile classes.

preTreatModes(processed_data, parameters, verbose = TRUE)

# S4 method for Binalysis
preTreatModes(processed_data, parameters, verbose = TRUE)

# S4 method for MetaboProfile
preTreatModes(processed_data, parameters, verbose = TRUE)

Arguments

processed_data
S4 object of class Binalysis or MetaboProfile

parameters
object of class AnalysisParameters containing pre-treatment parameters

verbose
console output

Value

S4 object of class Analysis

Examples

library(metabolyseR)

#> 
#> Attaching package: ‘metabolyseR’

#> The following object is masked from ‘package:stats’:
#> 
#>     anova

#> The following objects are masked from ‘package:base’:
#> 
#>     raw, split

## Retrieve file paths and sample information for example data
files <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')[1:2]

info <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')[1:2,]

## Detect spectral binning parameters
bp <- binneR::detectParameters(files)

## Perform spectral binning
analysis <- binneR::binneRlyse(files, 
                               info, 
                               parameters = bp)

#> binneR v2.5.3 Mon Apr 26 15:55:52 2021

#> ________________________________________________________________________________

#> Scans: 5:14 

#> ________________________________________________________________________________

#> Reading raw data

#> Gathering bins

#> Removing single scan events

#> Averaging intensities across scans

#> Calculating bin metrics

#> Calculating accurate m/z

#> Building intensity matrix

#> Gathering file headers

#> 

#> Completed! [2.7S]

## Declare pre-treatment parameters
pre_treatment_parameters <- analysisParameters('pre-treatment')
parameters(pre_treatment_parameters,
                        'pre-treatment') <- preTreatmentParameters(
 list(
   occupancyFilter = 'maximum',
   impute = 'all',
   transform = 'TICnorm'
 )
)
changeParameter(pre_treatment_parameters,'parallel') <- 'no'

## Perform pre-treatment
pre_treated_data <- preTreatModes(analysis,
                                  pre_treatment_parameters)

#> 
#> metabolyseR v0.14.1 Mon Apr 26 15:55:54 2021
#> ________________________________________________________________________________
#> Parameters:
#> pre-treatment
#> 	occupancyFilter
#> 		maximum
#> 			cls = class
#> 			occupancy = 2/3
#> 	impute
#> 		all
#> 			occupancy = 2/3
#> 			parallel = no
#> 			seed = 1234
#> 	transform
#> 		TICnorm
#> 
#> ________________________________________________________________________________
#> 
#> Negative mode…Negative mode 		✓ [0.7S]
#> Positive mode…Positive mode 		✓ [0.8S]
#> ________________________________________________________________________________
#> 
#> Complete! [1.6S]
#>