Reduce an analysis by removing isotopic, adduct or unknown features

Reduce an analysis by removing isotopic, adduct or unknown features.

reduce(x, isotopes = TRUE, adducts = TRUE, unknowns = FALSE)

# S4 method for Analysis
reduce(x, isotopes = TRUE, adducts = TRUE, unknowns = FALSE)

# S4 method for AnalysisData
reduce(x, isotopes = TRUE, adducts = TRUE, unknowns = FALSE)

Arguments

x
S4 class of class Analysis or AnalysisData

isotopes
TRUE/FALSE remove isotopic features.

adducts
TRUE/FALSE remove multiple adduct features.

unknowns
TRUE/FALSE remove unassigned features.

Value

S4 object of class Analysis or AnalysisData with features reduced

Details

Isotope and adduct features are filtered based on the maximum intensity peak for each molecular formulas.

Examples

library(MFassign)

#> Loading required package: ggraph

#> Loading required package: ggplot2

## Assign molecular formulas
p <- assignmentParameters('FIE')

assignment <- assignMFs(peakData,p)

#> 
#> MFassign v0.7.10 Mon May 10 15:02:15 2021

#> ________________________________________________________________________________

#> Assignment Parameters: 
#> 
#> 	 Technique:		 FIE 
#> 	 Max M:			 1000 
#> 	 Max MF score:		 5 
#> 	 PPM threshold:		 5 
#> 	 Relationship limit:	 0.001 
#> 
#> 	 Adducts: 
#> 	 n: [M-H]1-, [M+Cl]1-, [M+K-2H]1-, [M-2H]2-, [M+Cl37]1-, [2M-H]1-
#> 	 p: [M+H]1+, [M+K]1+, [M+Na]1+, [M+K41]1+, [M+NH4]1+, [M+2H]2+, [2M+H]1+ 
#> 	 Isotopes: 13C, 18O, 13C2 
#> 	 Transformations: M - [O] + [NH2], M - [OH] + [NH2], M + [H2], M - [H2] + [O], M - [H] + [CH3], M - [H] + [NH2], M - [H] + [OH], M + [H2O], M - [H3] + [H2O], M - [H] + [CHO2], M - [H] + [SO3], M - [H] + [PO3H2]

#> ________________________________________________________________________________

#> No. m/z:	9

#> Calculating correlations …

#> Calculating correlations 	✓ [24 correlations] [0.1S]

#> Filtering correlations …

#> Filtering correlations 		✓ [24 correlations] [0S]

#> Preparing correlations …

#> Preparing correlations 		✓ [0S]

#> Calculating relationships …

#> Calculating relationships 	✓ [8.2S]

#> Adduct & isotope assignment …

#> Adduct & isotope assignment 	✓ [13.7S]

#> Transformation assignment iteration 1 …

#> Transformation assignment iteration 1 	✓ [0S]

#> ________________________________________________________________________________

#> 
#> Complete! [22.1S]

## Retrieve assigned data
assigned_data <- metabolyseR::analysisData(
 assignedData(assignment),
 tibble::tibble(sample = seq_len(nrow(peakData)))
 )

reduced_data <- metaboMisc::reduce(assigned_data)

reduced_data

#> 
#> AnalysisData object containing:
#> 
#> Samples: 60 
#> Features: 1 
#> Info: 1 
#>